Fixed incorrect dwMaxFileCount value for SFileCreateArchive
So, took alittle awhile to find out why MPQs had an additional 8mb of
garbage data at the end of them compared to Ladik's editor. I thought
maybe there was a more siginificant fault related to writing the MPQ. In
was just a result of SFileCreateArchive in Stormsharplib always passing
int.MaxValue even if you provided your own count.

This fix address this.