Added MPQ Compression support via AddFileFromDiskWithCompression
Internally it calls the original unreferenced stormlib binding
SFileAddFileEx. With SFileAddFileFlags.MPQ_FILE_COMPRESS set and the
provided MpqCompressionTypeFlags from the method parameter list.

This is a pretty important feature for production. Without this feature
it is not reasonable to use this Stormlib library to construct MPQs, the
sizes will be too large.